% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{remove_custom_punctuation}
\alias{remove_custom_punctuation}
\title{Remove Custom Punctuation}
\usage{
remove_custom_punctuation(x)
}
\arguments{
\item{x}{A character vector of text}
}
\value{
A character vector with apostrophes and other punctuation removed
}
\description{
Remove Custom Punctuation
}
