% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{remove_custom_punctuation}
\alias{remove_custom_punctuation}
\title{Wordcloud That Displays the Most Frequently Used Word}
\usage{
remove_custom_punctuation(x)
}
\arguments{
\item{x}{The input string we want to process, like removing punctuation}
}
\value{
The output from \code{\link{wordcloud()}}
}
\description{
Wordcloud That Displays the Most Frequently Used Word
}
